validate oke

+++ src/proj_tt.Application/Products/Dto/ProductDto.cs
@@ -4,11 +4,11 @@ using Abp.AutoMapper;
 namespace proj_tt.Products.Dto
 {
     [AutoMapFrom(typeof(Product))]
-    public class ProductDto:AuditedEntityDto
+    public class ProductDto : AuditedEntityDto
     {
         public string Name { get; set; }
 
-        public float Price { get; set; }
+        public decimal Price { get; set; }
 
         public string ImageUrl { get; set; }
         public int Discount { get; set; }
